+++ NEWS.md
@@ -1,7 +1,25 @@
-## version 0.0.1.9000
+## version 0.2.0
 
 ---
 
+- api function is now set during api startup and not during runtime
+- fix saving of resutls into DB setup
+- adjust DB setup with name and init
+- fix adding of jobs
+- adding utils package as dependency
+- add DB updates for status
+- update .lintr options
+- fix id type to numeric
+- adjust logging
+- refactor DB setup and tests
+- add lintr and withr
+
+## version 0.1.0
+
+---
+
+- add RSQLite database to store queue and results
+- refactor background process
 - update documentation
 - fix bug with number of workers
 - fix typos